Move component to file
