+++ client/App.tsx
@@ -18,10 +18,7 @@ import {
   StatBlockEditorProps
 } from "./StatBlockEditor/StatBlockEditor";
 import { SpellEditor, SpellEditorProps } from "./StatBlockEditor/SpellEditor";
-import { InitiativeList } from "./InitiativeList/InitiativeList";
-import { CommandContext } from "./InitiativeList/CommandContext";
 import { useCallback } from "react";
-import { TagState } from "../common/CombatantState";
 import { PendingPrompts } from "./Prompts/PendingPrompts";
 import { CombatFooter } from "./CombatFooter/CombatFooter";
 import { SelectedCombatants } from "./SelectedCombatants";
@@ -29,6 +26,7 @@ import { VerticalResizer } from "./VerticalResizer";
 import { useStoreBackedState } from "./Utility/useStoreBackedState";
 import { Store } from "./Utility/Store";
 import { LegacySynchronousLocalStore } from "./Utility/LegacySynchronousLocalStore";
+import { InitiativeListHost } from "./InitiativeListHost";
 
 /*
  * This file is new as of 05/2020. Most of the logic was extracted from TrackerViewModel.
@@ -286,73 +284,5 @@ function ToolbarHost(props: { tracker: TrackerViewModel }) {
       width={toolbarWide ? "wide" : "narrow"}
       showCombatantCommands={combatantSelected}
     />
-  );
-}
-
-function InitiativeListHost(props: { tracker: TrackerViewModel }) {
-  const { tracker } = props;
-
-  const encounterState = useSubscription(
-    tracker.Encounter.ObservableEncounterState
-  );
-  const selectedCombatantIds = useSubscription(
-    tracker.CombatantCommander.SelectedCombatants
-  ).map(c => c.Combatant.Id);
-  const combatantCountsByName = useSubscription(
-    tracker.Encounter.CombatantCountsByName
-  );
-  const combatantViewModels = useSubscription(tracker.CombatantViewModels);
-
-  const selectCombatantById = useCallback(
-    (combatantId: string, appendSelection: boolean) => {
-      const selectedViewModel = combatantViewModels.find(
-        c => c.Combatant.Id == combatantId
-      );
-
-      if (selectedViewModel !== undefined) {
-        tracker.CombatantCommander.Select(selectedViewModel, appendSelection);
-      }
-    },
-    [tracker, combatantViewModels]
-  );
-
-  const removeCombatantTag = useCallback(
-    (combatantId: string, tagState: TagState) => {
-      const combatantViewModel = combatantViewModels.find(
-        c => c.Combatant.Id == combatantId
-      );
-      combatantViewModel?.RemoveTagByState(tagState);
-    },
-    [tracker, combatantViewModels]
-  );
-
-  const applyDamageToCombatant = useCallback(
-    (combatantId: string) => {
-      const combatantViewModel = combatantViewModels.find(
-        c => c.Combatant.Id == combatantId
-      );
-
-      if (combatantViewModel !== undefined) {
-        tracker.CombatantCommander.EditSingleCombatantHP(combatantViewModel);
-      }
-    },
-    [tracker, combatantViewModels]
-  );
-
-  return (
-    <CommandContext.Provider
-      value={{
-        SelectCombatant: selectCombatantById,
-        RemoveTagFromCombatant: removeCombatantTag,
-        ApplyDamageToCombatant: applyDamageToCombatant,
-        CombatantCommands: tracker.CombatantCommander.Commands
-      }}
-    >
-      <InitiativeList
-        encounterState={encounterState}
-        selectedCombatantIds={selectedCombatantIds}
-        combatantCountsByName={combatantCountsByName}
-      />
-    </CommandContext.Provider>
   );
 }
